+++ src/views/common/Welcome.vue
@@ -24,11 +24,6 @@
     </div>
 </template>
 <script lang="ts">
-    import UserManagement from "@/api/UserManagement";
-    import CourseManagement from "@/api/CourseManagement";
-    import AuthenticationManagement from "@/api/AuthenticationManagement";
-    import HyperledgerManagement from "@/api/HyperledgerManagement";
-    import HyperledgerCourseManagement from "@/api/HyperledgerCourseManagement";
     import { useStore } from "@/store/store";
     import { ref, onBeforeMount } from "vue";
     import LoadingSpinner from "@/components/loading/Spinner.vue";