+++ tests/unit/api/versionEndpoints.spec.ts
@@ -7,8 +7,6 @@ import Course from "@/api/api_models/course_management/Course";
 import { Language } from "@/entities/Language";
 import { CourseType } from "@/entities/CourseType";
 import AuthenticationManagement from "@/api/AuthenticationManagement";
-import HyperledgerManagement from "@/api/HyperledgerManagement";
-import HyperledgerCourseManagement from "@/api/HyperledgerCourseManagement";
 import MatriculationManagement from "@/api/MatriculationManagement";
 
 jest.useFakeTimers();
@@ -27,18 +25,6 @@ test("authentication service version", async () => {
 
 test("user service version", async () => {
     const version = await UserManagement.getVersion();
-
-    expect(version).not.toEqual("unavailable");
-});
-
-test("hyperledger management service version", async () => {
-    const version = await HyperledgerManagement.getVersion();
-
-    expect(version).not.toEqual("unavailable");
-});
-
-test("hyperledger course management service version", async () => {
-    const version = await HyperledgerCourseManagement.getVersion();
 
     expect(version).not.toEqual("unavailable");
 });